Alteracao dos modelos de vista com ratings de double para int.

+++ STrackerServer/STrackerServer/Controllers/EpisodesWebController.cs
@@ -152,7 +152,7 @@ namespace STrackerServer.Controllers
                 Poster = episode.Poster,
                 TvShowName = tvshow.Name,
                 Date = episode.Date,
-                Rating = episodeRating.Average,
+                Rating = (int)episodeRating.Average,
                 IsSubscribed = isSubscribed,
                 Watched = watched,
                 AsAired = !(DateTime.Compare(episodeDate, DateTime.Now) > 0),
